Extract heater FSM to a separate file
